remove old components
