--- src/components/LinkerPagev2/index.ts
@@ -1,2 +1,0 @@
-import LinkerPage from './LinkerPage.container'
-export default LinkerPage
